fixed buttons and display
