+++ project/toy/switches.h
@@ -4,7 +4,7 @@
 #define SW1 BIT0
 #define SW2 BIT1
 #define SW3 BIT2
-#define SW4 BIT4
+#define SW4 BIT3
 
 #define SWITCHES (SW1 | SW2 | SW3 | SW4)
